moved definition of withoutRef function and removed Form tsx file
